requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
playwright>=1.40.0
fastapi>=0.104.1
uvicorn[standard]>=0.23.2
pydantic>=2.4.2
google-genai>=0.1.0
python-dotenv>=1.0.0
